sctp-tests: add -T opt to allow users to pass topo opt
This patch is to move def topo opt from testcase to topo file,
and allow users to change topo opt when running a case. For
team, the usage is like:

  # ./sctp-tests -e testcase/misc/demo/test_with_team.sh \
        -T "st_team_opt='{\"runner\":{\"name\":\"activebackup\"}'}" run

If in test cases, Topo_opt is set, topos will use it. If It is not,
then try to use -T param. If -T param is not set, use the def_opt.

Signed-off-by: Xin Long <lucien.xin@gmail.com>
Signed-off-by: Marcelo Ricardo Leitner <marcelo.leitner@gmail.com>