testcase: add a performance case sctphashtable
This patch is to add sctphashtable on 3 topos: phost, netns and team.
It is used to test sctp transport rhashtable.

Note that we don't reused the codes between phost and netns even if
these two topos have so many codes in common. Because phost and netns
are really different, sharing codes will make things worse in the
future.

But different subtopos for netns, sharing codes is welcomed, as well
as for phost.

Signed-off-by: Xin Long <lucien.xin@gmail.com>
Signed-off-by: Marcelo Ricardo Leitner <marcelo.leitner@gmail.com>